Create company_bias.md
let's think 4 different development style:

Linus: do not write code unless necessary. user is smart enough.
Microsoft: if restart solves the problem, it is NOT a problem.
Google: We don't want to do evil things, but we never knows. So we do things we like.
Apple: User is stupid, we MUST protect them.